backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
requests

# OCR
easyocr==1.7.1
opencv-python-headless
Pillow
pdf2image

# Documents
python-docx
python-pptx
PyPDF2

# MongoDB
motor
pymongo

# Utils
python-dateutil
tqdm
